refactor: use allow function
