+++ src/Bookmarks/component.tsx
@@ -1,4 +1,4 @@
-import { FC, useEffect, useState } from 'react'
+import { FC } from 'react'
 import {
   Bookmark,
   BookmarkFolder,
@@ -6,9 +6,9 @@ import {
   collapseBookmarkFolder,
   expandBookmarkFolder,
 } from './model'
-import { subscribeBookmarks } from './repository'
 
 import './component.css'
+import { useBookmarkFolders, useLocalStorage } from './hook'
 
 export const BookmarkFolders: FC = () => {
   const bookmarkFolders = useBookmarkFolders()
@@ -98,56 +98,4 @@ const BookmarkComponent: FC<BookmarkComponentProps> = ({ bookmark }) => {
       </div>
     </div>
   )
-}
-
-function useBookmarkFolders() {
-  const [bookmarkFolders, setBookmarkFolders] = useState<BookmarkFolder[]>([])
-  useEffect(() => {
-    const subscription = subscribeBookmarks((bookmarkFolders) => {
-      setBookmarkFolders(bookmarkFolders)
-    })
-    subscription.refresh()
-    return () => {
-      subscription.unsubscribe()
-    }
-  }, [])
-  return bookmarkFolders
-}
-
-function useLocalStorage<T>(localStorageKey: string, initialValue: T): [T, (value: T) => void] {
-  const [storedValue, setStoredValue] = useState<T>(() => {
-    const value = localStorage.getItem(localStorageKey)
-    if (value === null) {
-      return initialValue
-    }
-    try {
-      return JSON.parse(value) as T
-    } catch {
-      return initialValue
-    }
-  })
-
-  useEffect(() => {
-    function handleStorageEvent(e: StorageEvent) {
-      if (e.storageArea === localStorage && e.key === localStorageKey && e.newValue !== null) {
-        try {
-          setStoredValue(JSON.parse(e.newValue) as T)
-        } catch {
-          setStoredValue(initialValue)
-        }
-      }
-    }
-    window.addEventListener('storage', handleStorageEvent)
-    return () => {
-      window.removeEventListener('storage', handleStorageEvent)
-    }
-  }, [setStoredValue, localStorageKey, initialValue])
-
-  return [
-    storedValue,
-    (value: T) => {
-      setStoredValue(value)
-      localStorage.setItem(localStorageKey, JSON.stringify(value))
-    },
-  ]
 }
