+++ src/TopSites/repository.ts
@@ -1,5 +1,3 @@
 import { TopSite } from './model'
 
-export function getTopSites(callback: (topSites: TopSite[]) => void) {
-  chrome.topSites.get(callback)
-}
+export const getTopSites = (callback: (topSites: TopSite[]) => void) => chrome.topSites.get(callback)
